requirements.txt
Django==4.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
djangorestframework-jsonapi==6.1.0
django-filter==24.3
Pillow==9.4.0
psycopg2-binary==2.9.5
mysqlclient==2.1.1 
gunicorn==20.1.0  
django-cors-headers==4.1.0 
django-environ==0.11.2
drf-yasg==1.21.7
stripe==10.8.0
celery==5.4.0
redis==5.0.7
django-celery-beat==2.6.0
pytest==8.3.2
pytest-cov==5.0.0
pytest-django==4.8.0
flake8==6.0.0
coverage==7.6.1
python-telegram-bot==21.3
django-celery-results==2.5.1
python-decouple==3.8